+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 ==================
 
   * Bottom bar with some information about tasks
+  * Added configuration Menu
   * Scroll now only visible inside the suite
   * Window title changed to: Gaucho Task Launcher
-  * Config now using electron-config package
+  * Configuration data is now stored using electron-config package
   * Minimum window width increased to avoid horizontal scroll
   * Overall style slightly improved
   * Fallback to default config if file is corrupted